+++ packages/client-generator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @redocly/client-generator
 
+## 0.4.5
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.51.1.
+
+## 0.4.4
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.51.0.
+
 ## 0.4.3
 
 ### Patch Changes